+++ tests/testthat/test-r.R
@@ -14,10 +14,7 @@ test_that("we can use R CMD build to build a package", {
 
   package <- "sample.package"
   pkgdir <- file.path(testdir, package)
-  ensure_directory(pkgdir)
-
-  data <- list(Package = package, Type = "Package", Version = "0.1.0")
-  renv_dcf_write(data, file = file.path(pkgdir, "DESCRIPTION"))
+  renv_package_skeleton(package, pkgdir)
   expect_equal(renv_project_type(pkgdir), "package")
 
   tarball <- r_cmd_build(package, pkgdir)